hubspot-slack-digest-agent/requirements.txt
python-dotenv>=1.0.0
scalekit-sdk-python>=1.0.0
